Fix JSON serialization error for numpy float32 in MQTT publishing
Convert numpy float32 to native Python float in BarkEvent.to_dict()
to ensure proper JSON serialization when publishing to MQTT broker.

The round() function preserves numpy types, so explicit conversion
with float() is required for JSON compatibility.

Fixes: Error publishing event: Object of type float32 is not JSON serializable

+++ bark_detector/detector.py
@@ -34,7 +34,7 @@ class BarkEvent:
         """Convert to dictionary for JSON serialization."""
         return {
             "timestamp": self.timestamp,
-            "peak_db": round(self.peak_db, 1),
+            "peak_db": float(round(self.peak_db, 1)),
             "device": self.device
         }
